add ProjectClient: - fix project query to project service instead of metadata service
