+++ app/clients/__init__.py
@@ -1,0 +1,2 @@
+from .metadata import MetadataClient  # noqa
+from .project import ProjectClient  # noqa
