+++ app/clients/base.py
@@ -1,0 +1,34 @@
+# Copyright (C) 2022 Indoc Research
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+from typing import Any
+from typing import Dict
+
+import httpx
+
+from app.config import ConfigClass
+
+
+class BaseClient:
+
+    BASE_URL = ConfigClass.PROJECT_SERVICE
+
+    @classmethod
+    async def get(cls, url: str, params: Dict[str, Any] = None) -> Dict[str, Any]:
+        async with httpx.AsyncClient() as client:
+            response = await client.get(url, params=params)
+        response.raise_for_status()
+
+        return response.json()['result']
