+++ app/clients/metadata.py
@@ -20,18 +20,12 @@ import httpx
 
 from app.config import ConfigClass
 
+from .base import BaseClient
 
-class MetadataClient:
+
+class MetadataClient(BaseClient):
 
     BASE_URL = ConfigClass.METADATA_SERVICE
-
-    @classmethod
-    async def get(cls, url: str, params: Dict[str, Any] = None) -> Dict[str, Any]:
-        async with httpx.AsyncClient() as client:
-            response = await client.get(url, params=params)
-        response.raise_for_status()
-
-        return response.json()['result']
 
     @classmethod
     async def get_objects(cls, code: str) -> Dict[str, Any]:
