+++ tests/conftest.py
@@ -41,6 +41,7 @@ environ['ENTITYINFO_SERVICE'] = 'http://ENTITYINFO_SERVICE'
 environ['ELASTIC_SEARCH_SERVICE'] = 'http://ELASTIC_SEARCH_SERVICE'
 environ['SEND_MESSAGE_URL'] = 'http://SEND_MESSAGE_URL'
 environ['METADATA_SERVICE'] = 'http://METADATA_SERVICE'
+environ['PROJECT_SERVICE'] = 'http://PROJECT_SERVICE'
 
 environ['gm_queue_endpoint'] = 'http://gm_queue_endpoint'
 environ['gm_username'] = 'gm_username'
