+++ tests/dataset_file/test_import_files_list.py
@@ -43,7 +43,7 @@ async def test_import_files_from_source_list_should_return_200(client, httpx_moc
     }
     httpx_mock.add_response(
         method='GET',
-        url=f'http://metadata_service/v1/item/{source_project}/',
+        url=f'http://project_service/v1/projects/{source_project}/',
         json={'result': {'code': 'project_code'}},
     )
     httpx_mock.add_response(
@@ -153,7 +153,7 @@ async def test_05_test_import_duplicate(client, httpx_mock, dataset):
     }
     httpx_mock.add_response(
         method='GET',
-        url=f'http://metadata_service/v1/item/{source_project}/',
+        url=f'http://project_service/v1/projects/{source_project}/',
         json={'result': {'code': 'project_code'}},
     )
     httpx_mock.add_response(
